Initial commit for creation of rescue OS image on USB.
This uses the already obsolete Ubuntu Xenial (16.04) release because
it is the closest to synology.

As such the generated image SHOULD NOT be used for day to day use
the purpose is to have the ability to read your btrfs data if all
else is lost e.g. fact it exists should provide sufficient karma
that you don't ever need it.