Add support for building Live CD ISO
By adding setup required to create EFI and ISOLINUX images to the
base rootfs (casper, isolinux) we can fairly trivially convert the
built rootfs image to a live CD with Xorisso.  These images are hybrid
EFI images so they could in theory be crammed on a USB but that is
easier with direct generation.  The main point is to have a bootable
DVD in the future.

Since DVD has enough room for full desktop added a couple other
improvements:
 - Full Bionic desktop installation
 - Ability to build with loop devices so DVD is possible without media.

Finally moved hooks to hooks/all with links in hooks/enabled.  This
makes it simpler to customize without changing git repo and also
gets rid of errors due to editor droppings.